requirements.txt
uuid
python-dotenv
google-generativeai